[ADDED] New Category Policy

+++ app/Http/Controllers/Auth/LoginController.php
@@ -40,9 +40,11 @@ class LoginController extends Controller
         // if(OrderItem::truncate())
         //     Order::truncate();
         // dd($result);
-        $order = Order::find(session('order_id'));
-        $order->orderItems()->delete();
-        $order->delete();
+        if(session('order_id')){
+            $order = Order::find(session('order_id'));
+            $order->orderItems()->delete();
+            $order->delete();
+        }
         
         Auth::logout();
         $request->session()->invalidate();
